Further refactoring:
- Move `TagVar` and `TagVars` to `picard/tags/tagvar.py`
- Move `picard/util/preservedtags.py` to `picard/tags/preservedtags.py`
- Move common tags functions to `picard/tags/__init__.py`
- Rename `test_util_tags.py` to `test_tags.py`
- Update `import` statments in affected files including tests
- Update affected `mock` statements in tests
- Move and combine `test_display_tag_name()` test from `test_utile.py`
into `test_tags.py`
