+++ picard/tags/__init__.py
@@ -31,7 +31,6 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
-
 import re
 
 from picard.const.tags import ALL_TAGS
