+++ picard/ui/widgets/taglisteditor.py
@@ -22,7 +22,7 @@
 
 from PyQt6 import QtWidgets
 
-from picard.util.tags import tag_names
+from picard.tags import tag_names
 
 from picard.ui.forms.ui_widget_taglisteditor import Ui_TagListEditor
 from picard.ui.widgets.editablelistview import (
